Alter calixis ConfigGrid to use WeightedRegions instead of storing the weights in the ConfigRegion.
