+++ calixis/plan/models.py
@@ -5,3 +5,4 @@ from .submodels.inspiration import *
 from .submodels.job import *
 from .submodels.perterbation import *
 from .submodels.roll import *
+from .submodels.weighted import *
